Add o2 service for consistency when wallet is connected
