+++ docs/src/lib/components/examples/connect-wallet/ReactConnectWallet.md
@@ -6,26 +6,116 @@ Import the libraries and any wallets you would like to use. For this example, we
 ```js title="App.tsx"|copy
 import { Web3OnboardProvider, init } from '@web3-onboard/react'
 import injectedModule from '@web3-onboard/injected-wallets'
+import fortmaticModule from '@web3-onboard/fortmatic'
+import gnosisModule from '@web3-onboard/gnosis'
+import injectedModule from '@web3-onboard/injected-wallets'
+import keepkeyModule from '@web3-onboard/keepkey'
+import keystoneModule from '@web3-onboard/keystone'
+import ledgerModule from '@web3-onboard/ledger'
+import portisModule from '@web3-onboard/portis'
+import torusModule from '@web3-onboard/torus'
+import trezorModule from '@web3-onboard/trezor'
+import walletConnectModule from '@web3-onboard/walletconnect'
+import coinbaseModule from '@web3-onboard/coinbase'
+import magicModule from '@web3-onboard/magic'
+import web3authModule from '@web3-onboard/web3auth'
+import dcentModule from '@web3-onboard/dcent'
+import sequenceModule from '@web3-onboard/sequence'
+import tallyHoModule from '@web3-onboard/tallyho'
 
 const INFURA_KEY = ''
 
-const ethereumRopsten = {
-  id: '0x3',
-  token: 'rETH',
-  label: 'Ethereum Ropsten',
-  rpcUrl: `https://ropsten.infura.io/v3/${INFURA_KEY}`
+const injected = injectedModule()
+const coinbase = coinbaseModule()
+const dcent = dcentModule()
+const walletConnect = walletConnectModule()
+
+const portis = portisModule({
+  apiKey: 'apiKey'
+})
+
+const fortmatic = fortmaticModule({
+  apiKey: 'apiKey'
+})
+
+const ledger = ledgerModule()
+const keystone = keystoneModule()
+const keepkey = keepkeyModule()
+const gnosis = gnosisModule()
+const sequence = sequenceModule()
+const tally = tallyModule()
+
+const trezorOptions = {
+  email: 'test@test.com',
+  appUrl: 'https://www.blocknative.com'
 }
 
-const polygonMainnet = {
-  id: '0x89',
-  token: 'MATIC',
-  label: 'Polygon',
-  rpcUrl: 'https://matic-mainnet.chainstacklabs.com'
-}
+const trezor = trezorModule(trezorOptions)
 
-const chains = [ethereumRopsten, polygonMainnet]
+const magic = magicModule({
+  apiKey: 'apiKey'
+})
 
-const wallets = [injectedModule()]
+const enkrypt = enkryptModule()
+const mewWallet = mewWalletModule()
+
+const wallets = [
+  keepkey,
+  sequence,
+  injected,
+  tally,
+  ledger,
+  coinbase,
+  dcent,
+  trezor,
+  walletConnect,
+  enkrypt,
+  mewWallet,
+  gnosis,
+  magic,
+  fortmatic,
+  keystone,
+  portis
+]
+
+const chains = [
+  {
+    id: '0x1',
+    token: 'ETH',
+    label: 'Ethereum Mainnet',
+    rpcUrl: `https://mainnet.infura.io/v3/${INFURA_ID}`
+  },
+  {
+    id: '0x5',
+    token: 'ETH',
+    label: 'Goerli',
+    rpcUrl: `https://goerli.infura.io/v3/${INFURA_ID}`
+  },
+  {
+    id: '0x13881',
+    token: 'MATIC',
+    label: 'Polygon - Mumbai',
+    rpcUrl: 'https://matic-mumbai.chainstacklabs.com'
+  },
+  {
+    id: '0x38',
+    token: 'BNB',
+    label: 'Binance',
+    rpcUrl: 'https://bsc-dataseed.binance.org/'
+  },
+  {
+    id: '0xA',
+    token: 'OETH',
+    label: 'Optimism',
+    rpcUrl: 'https://mainnet.optimism.io'
+  },
+  {
+    id: '0xA4B1',
+    token: 'ARB-ETH',
+    label: 'Arbitrum',
+    rpcUrl: 'https://rpc.ankr.com/arbitrum'
+  }
+]
 
 const appMetadata = {
   name: 'Connect Wallet Example',
