+++ docs/src/lib/components/examples/connect-wallet/SvelteConnectWallet.md
@@ -5,10 +5,77 @@ Import the libraries and any wallets you would like to use. For this example, we
 ```js title="onboard.js"|copy
 import Onboard from '@web3-onboard/core'
 import injectedModule from '@web3-onboard/injected-wallets'
+import fortmaticModule from '@web3-onboard/fortmatic'
+import gnosisModule from '@web3-onboard/gnosis'
+import injectedModule from '@web3-onboard/injected-wallets'
+import keepkeyModule from '@web3-onboard/keepkey'
+import keystoneModule from '@web3-onboard/keystone'
+import ledgerModule from '@web3-onboard/ledger'
+import portisModule from '@web3-onboard/portis'
+import torusModule from '@web3-onboard/torus'
+import trezorModule from '@web3-onboard/trezor'
+import walletConnectModule from '@web3-onboard/walletconnect'
+import coinbaseModule from '@web3-onboard/coinbase'
+import magicModule from '@web3-onboard/magic'
+import web3authModule from '@web3-onboard/web3auth'
+import dcentModule from '@web3-onboard/dcent'
+import sequenceModule from '@web3-onboard/sequence'
+import tallyHoModule from '@web3-onboard/tallyho'
 
 const INFURA_KEY = ''
 
-const wallets = [injectedModule()]
+const injected = injectedModule()
+const coinbase = coinbaseModule()
+const dcent = dcentModule()
+const walletConnect = walletConnectModule()
+
+const portis = portisModule({
+  apiKey: 'apiKey'
+})
+
+const fortmatic = fortmaticModule({
+  apiKey: 'apiKey'
+})
+
+const ledger = ledgerModule()
+const keystone = keystoneModule()
+const keepkey = keepkeyModule()
+const gnosis = gnosisModule()
+const sequence = sequenceModule()
+const tally = tallyModule()
+
+const trezorOptions = {
+  email: 'test@test.com',
+  appUrl: 'https://www.blocknative.com'
+}
+
+const trezor = trezorModule(trezorOptions)
+
+const magic = magicModule({
+  apiKey: 'apiKey'
+})
+
+const enkrypt = enkryptModule()
+const mewWallet = mewWalletModule()
+
+const wallets = [
+  keepkey,
+  sequence,
+  injected,
+  tally,
+  ledger,
+  coinbase,
+  dcent,
+  trezor,
+  walletConnect,
+  enkrypt,
+  mewWallet,
+  gnosis,
+  magic,
+  fortmatic,
+  keystone,
+  portis
+]
 
 const chains = [
   {
@@ -18,10 +85,34 @@ const chains = [
     rpcUrl: `https://mainnet.infura.io/v3/${INFURA_ID}`
   },
   {
-    id: '0x89',
+    id: '0x5',
+    token: 'ETH',
+    label: 'Goerli',
+    rpcUrl: `https://goerli.infura.io/v3/${INFURA_ID}`
+  },
+  {
+    id: '0x13881',
     token: 'MATIC',
-    label: 'Polygon',
-    rpcUrl: 'https://matic-mainnet.chainstacklabs.com'
+    label: 'Polygon - Mumbai',
+    rpcUrl: 'https://matic-mumbai.chainstacklabs.com'
+  },
+  {
+    id: '0x38',
+    token: 'BNB',
+    label: 'Binance',
+    rpcUrl: 'https://bsc-dataseed.binance.org/'
+  },
+  {
+    id: '0xA',
+    token: 'OETH',
+    label: 'Optimism',
+    rpcUrl: 'https://mainnet.optimism.io'
+  },
+  {
+    id: '0xA4B1',
+    token: 'ARB-ETH',
+    label: 'Arbitrum',
+    rpcUrl: 'https://rpc.ankr.com/arbitrum'
   }
 ]
 
@@ -38,7 +129,7 @@ const appMetadata = {
 const onboard = Onboard({
   wallets,
   chains,
-  appMetadata,
+  appMetadata
 })
 
 export default onboard
@@ -50,15 +141,11 @@ In main `App.svelte` file we'll import our previously initialized web3-onboard i
 
 ```svelte title="App.svelte"|copy
 <script lang="js">
-import onboard from './onboard.js'
+  import onboard from './onboard.js'
 </script>
 
 <div>
-  <button
-    disabled={connecting}
-    onClick={connect}>
-    Connect
-  </button>
+  <button disabled={connecting} onClick={connect}> Connect </button>
 </div>
 ```
 
@@ -69,7 +156,7 @@ Now that we have our wallet connected, let's display some basic information, suc
 ```svelte title="App.svelte"|copy
 <script lang="js">
   import onboard from './onboard.js'
-  
+
   // Subscribe to wallet updates
   const wallets$ = onboard.state.select('wallets')
 
@@ -82,7 +169,7 @@ Now that we have our wallet connected, let's display some basic information, suc
       address: connectedAccount?.address
     }
   : { address: connectedAccount?.address }
-  
+
   const connect = async () => {
     await onboard.connectWallet()
   }
